Compiled Arduino sketch
I created this sensor package for use in a project for the Sherman Centre for Digital Scholarship at McMaster University. The wiki for this whole project can be found here: https://macwiki.mcmaster.ca/posthumanism/index.php/Main/GreyhoundSensorRig
Note that this is my first real Arduino project, so the code is likely not as clean or elegant as some might hope! There are plenty of comments that hopefully make its functionality clear.

Using Arduino example code and test code found in the Seeed wiki, I compiled this sketch in order to use four Grove sensors together, using an Arduino Uno, a Grove base shield, and an Elecrow RTC Data Logger shield. This is the first complete version of this code. The sensors used in this package are:
- Grove 3-Axis Digital Accelerometer (±16g)
- Grove 3-Axis Analog Accelerometer 
- Grove Ear-Clip Heart Rate Sensor
- Grove Loudness Sensor